rename class and add namespace
